+++ json_serializable/lib/src/json_key_helpers.dart
@@ -2,7 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:analyzer/analyzer.dart';
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:analyzer/dart/constant/value.dart';
@@ -98,7 +97,7 @@ ConvertData _getFunctionName(
   var functionElement = type.element as FunctionElement;
 
   var positionalParams = functionElement.parameters
-      .where((pe) => pe.parameterKind == ParameterKind.REQUIRED)
+      .where((pe) => pe.isPositional)
       .toList();
 
   if (positionalParams.length != 1) {